if tmp bundle file exists, remove it

+++ collect_wordlists.sh
@@ -16,12 +16,9 @@ curl -o "${TARGET_DIR}"/danielmiessler-SecLists-graphql.txt "https://raw.githubu
 curl -o "${TARGET_DIR}"/kiterunner-swagger-wordlist.txt "https://wordlists-cdn.assetnote.io/data/kiterunner/swagger-wordlist.txt"
 curl -o "${TARGET_DIR}"/httparchive_apiroutes_2021_08_28.txt "https://wordlists-cdn.assetnote.io/./data/automated/httparchive_apiroutes_2021_08_28.txt"
 
-# remove current bundle
-rm -- "*.txt"
-
 # create new bundle
 cat -- "${TMP_DIR}/*.txt"  > "${TMP_BUNDLE}"
 cat -- "${TMP_BUNDLE}" | sort -n | uniq > "${TARGET_DIR}/wordlist_bundle_$(date +"%Y-%m-%d").txt"
 
 # cleanup
-rm -fv -- "${TMP_BUNDLE}"
+test -f "${TMP_BUNDLE}" && rm -fv -- "${TMP_BUNDLE}"
